Move navigation into separate folder
