+++ src/navigation/DashboardStack.js
@@ -1,5 +1,5 @@
 import { createStackNavigator } from 'react-navigation'
-import DashboardView from '../../screens/Dashboard'
+import DashboardView from '../screens/Dashboard'
 import { generateNavOptions } from './helpers'
 
 export default createStackNavigator({
