+++ src/navigation/DrawerNavigator.js
