+++ src/navigation/FamiliesStack.js
@@ -1,7 +1,7 @@
 import { createStackNavigator } from 'react-navigation'
-import FamiliesView from '../../screens/Families'
+import FamiliesView from '../screens/Families'
 import LifemapScreens from './LifemapScreens'
-import FamilyView from '../../screens/Family'
+import FamilyView from '../screens/Family'
 import { generateNavOptions } from './helpers'
 
 export default createStackNavigator({
