+++ src/navigation/LifemapScreens.js
@@ -1,22 +1,22 @@
 import React from 'react'
 import { Text, StyleSheet, Platform } from 'react-native'
-import TermsView from '../../screens/lifemap/Terms'
-import SocioEconomicQuestionView from '../../screens/lifemap/SocioEconomicQuestion'
-import FinalView from '../../screens/lifemap/Final'
-import FamilyParticipantView from '../../screens/lifemap/FamilyParticipant'
-import FamilyMembersNamesView from '../../screens/lifemap/FamilyMembersNames'
-import QuestionView from '../../screens/lifemap/Question'
-import BeginLifemapView from '../../screens/lifemap/BeginLifemap'
-import LocationView from '../../screens/lifemap/Location'
-import SkippedView from '../../screens/lifemap/Skipped'
-import OverviewView from '../../screens/lifemap/Overview'
-import AddPriorityView from '../../screens/lifemap/AddPriority'
-import AddAchievementView from '../../screens/lifemap/AddAchievement'
-import FamilyMemberView from '../../screens/lifemap/FamilyMember'
-import i18n from '../../i18n'
+import TermsView from '../screens/lifemap/Terms'
+import SocioEconomicQuestionView from '../screens/lifemap/SocioEconomicQuestion'
+import FinalView from '../screens/lifemap/Final'
+import FamilyParticipantView from '../screens/lifemap/FamilyParticipant'
+import FamilyMembersNamesView from '../screens/lifemap/FamilyMembersNames'
+import QuestionView from '../screens/lifemap/Question'
+import BeginLifemapView from '../screens/lifemap/BeginLifemap'
+import LocationView from '../screens/lifemap/Location'
+import SkippedView from '../screens/lifemap/Skipped'
+import OverviewView from '../screens/lifemap/Overview'
+import AddPriorityView from '../screens/lifemap/AddPriority'
+import AddAchievementView from '../screens/lifemap/AddAchievement'
+import FamilyMemberView from '../screens/lifemap/FamilyMember'
+import i18n from '../i18n'
 import { generateNavOptions } from './helpers'
 import CustomHeaderSurvey from './CustomHeaderSurvey'
-import colors from '../../theme.json'
+import colors from '../theme.json'
 
 // Reusable object for all screens related to a draft
 export default {
