+++ src/navigation/LogoutPopup.js
@@ -2,13 +2,13 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Text, StyleSheet, View, Platform } from 'react-native'
 import CommunityIcon from 'react-native-vector-icons/MaterialCommunityIcons'
-import Popup from '../Popup'
-import Checkbox from '../Checkbox'
+import Popup from '../components/Popup'
+import Checkbox from '../components/Checkbox'
 import Icon from 'react-native-vector-icons/MaterialIcons'
-import Button from '../Button'
-import i18n from '../../i18n'
-import colors from '../../theme.json'
-import globalStyles from '../../globalStyles'
+import Button from '../components/Button'
+import i18n from '../i18n'
+import colors from '../theme.json'
+import globalStyles from '../globalStyles'
 
 export default class LogoutPopup extends Component {
   render() {
@@ -31,8 +31,7 @@ export default class LogoutPopup extends Component {
           <Icon name="close" size={20} />
         </View>
 
-        <View style={styles.modalContainer} 
-        accessibilityLiveRegion='polite'>
+        <View style={styles.modalContainer} accessibilityLiveRegion="polite">
           <View style={{ alignItems: 'center' }}>
             {!checkboxesVisible ? (
               <Icon
