+++ src/navigation/SyncStack.js
@@ -1,5 +1,5 @@
 import { createStackNavigator } from 'react-navigation'
-import SyncView from '../../screens/Sync'
+import SyncView from '../screens/Sync'
 import { generateNavOptions } from './helpers'
 
 export default createStackNavigator({
